DEV: tests- implement TestBaseClassWeb
