+++ tests/data/adminViews.php
@@ -1,0 +1,4 @@
+<?php
+return [
+    'login'=>'authentication/sa/login',
+];
